% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_custom_metadata}
\alias{add_custom_metadata}
\title{Add custom metadata}
\usage{
add_custom_metadata(dd = NULL, levels = NULL, sessions = NULL)
}
\arguments{
\item{dd}{data frame. Custom data dictionary.
Should have the same structure as the data dictionary provided by
the package.}

\item{levels}{data frame. Custom levels table.}

\item{sessions}{data frame. Custom sessions table.}
}
\value{
invisible \code{TRUE}.
}
\description{
This function allows users to add custom metadata
(data dictionary, levels table, sessions table) to the package environment.
This can be useful for users who want to use their own metadata instead of
the ones provided by the package, or for testing and development purposes.
}
\details{
The custom metadata will be stored in the package environment can be
accessed with any function that contains the "release" argument with
\code{release = "custom"}. For example,
\code{get_dd(study = "abcd", release = "custom")}.

The default value for \code{dd}, \code{levels}, and \code{sessions} is \code{NULL}.
If any of them is not \code{NULL}, it will be added to the package environment.
If the argument is \code{NULL}, the corresponding metadata will not be added.
}
\examples{
add_custom_metadata(
  dd = tibble::tibble(
    name = "var1",
    table_name = "table1",
    identifier_columns = "participant_id"
  )
)
get_dd(study = "abcd", release = "custom")
}
