% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bids_sidecar}
\alias{create_bids_sidecar}
\title{Create BIDS sidecar}
\usage{
create_bids_sidecar(
  data,
  study,
  release = "latest",
  var_coding = "values",
  metadata_description = "Dataset exported using NBDCtools",
  path_out = NULL,
  pretty = TRUE
)
}
\arguments{
\item{data}{tibble. The raw data or data with labels, see
\code{\link[=transf_label]{transf_label()}}. If the data is not labelled, the function will
try to label the data first.}

\item{study}{character. NBDC study (One of \code{"abcd"} or \code{"hbcd"})}

\item{release}{character. Release version (Default: \code{"latest"}).}

\item{var_coding}{character. the variable coding, one of "values", "labels".
If the data is processed with \code{\link[=transf_value_to_label]{transf_value_to_label()}}, please use "labels".}

\item{metadata_description}{string, the description of the metadata}

\item{path_out}{character. the path to the output file.
If \code{NULL}, the function will return the json object.}

\item{pretty}{logical. Whether to pretty print the json.}
}
\value{
the json object or the path to the json file
}
\description{
Creates a Brain Imaging Data Structure (BIDS) JSON sidecar file from the
metadata (data dictionary and levels table). Returns the JSON object or
writes it to a file.
}
\examples{
\dontrun{
data |> create_bids_sidecar()
data |> create_bids_sidecar(path_out = "data.json")
}
}
