% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bids_sidecar_metadata}
\alias{create_bids_sidecar_metadata}
\title{Create BIDS sidecar from metadata}
\usage{
create_bids_sidecar_metadata(
  dd,
  levels,
  vars = NULL,
  tables = NULL,
  metadata_description = "Dataset exported using NBDCtools",
  path_out = NULL,
  pretty = TRUE
)
}
\arguments{
\item{dd}{tibble, Data dictionary metadata, see \code{\link[=get_dd]{get_dd()}} or provide
a custom data dictionary. The custom data dictionary must have
same structure as the official data dictionaries.}

\item{levels}{tibble, Levels metadata corresponding to \code{dd},
see \code{\link[=get_levels]{get_levels()}}, or provide a custom levels table. The custom levels table
must have same structure as the official levels tables.}

\item{vars}{character vector,  variable names to include. \code{vars} and
\code{tables} cannot be both \code{NULL}.}

\item{tables}{character vector, table names to include.}

\item{metadata_description}{string, the description of the metadata}

\item{path_out}{character. the path to the output file.
If \code{NULL}, the function will return the json object.}

\item{pretty}{logical. Whether to pretty print the json.}
}
\value{
Either a JSON string (when \code{path_out} is \code{NULL}) or the output
file path (invisibly) after writing the JSON sidecar to disk.
}
\description{
Generates a Brain Imaging Data Structure (BIDS) JSON sidecar using
metadata tables (data dictionary and levels) without requiring the
underlying data.
}
\details{
\itemize{
\item If you have a labelled dataset, and want to create data specific
BIDS sidecar with variable levels from the data, please use
\code{\link[=create_bids_sidecar_data]{create_bids_sidecar_data()}}.
\item If you want to create a BIDS sidecar without the underlying data,
please use \code{\link[=create_bids_sidecar_metadata]{create_bids_sidecar_metadata()}}.
}
}
\examples{
\dontshow{if (requireNamespace("NBDCtoolsData", quietly = TRUE)) withAutoprint(\{ # examplesIf}
create_bids_sidecar_metadata(
  dd = get_dd("abcd"),
  levels = get_levels("abcd"),
  tables = c("ph_y_mctq")
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=create_bids_sidecar_data]{create_bids_sidecar_data()}}
}
