% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_sessions}
\alias{get_sessions}
\alias{get_sessions_abcd}
\alias{get_sessions_hbcd}
\title{Get sessions table}
\usage{
get_sessions(study, release = "latest")

get_sessions_abcd(...)

get_sessions_hbcd(...)
}
\arguments{
\item{study}{character. The study name. One of "abcd" or "hbcd".}

\item{release}{character. Release version (Default: \code{"latest"}).}

\item{...}{Additional arguments passed to the underlying
\code{\link[=get_sessions]{get_sessions()}} function.}
}
\value{
Data frame with the sessions table.
}
\description{
Retrieves the sessions table for a given study and release version. Wrapper
around \code{\link[=get_metadata]{get_metadata()}}.

In addition to the main \code{get_sessions()} function, there are two
study-specific variations:
\itemize{
\item \code{get_sessions_abcd()}: for the ABCD study.
\item \code{get_sessions_hbcd()}: for the HBCD study.
}

They have the same arguments as the \code{get_sessions()} function, except
that the \code{study} argument is set to the respective study by default, and
should not be set by the user.
}
\examples{
\dontshow{if (requireNamespace("NBDCtoolsData", quietly = TRUE)) withAutoprint(\{ # examplesIf}
get_sessions("abcd")

get_sessions("hbcd")

get_sessions_abcd(release = "6.0")

get_sessions_hbcd(release = "1.0")
\dontshow{\}) # examplesIf}
}
