% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url_deap}
\alias{get_url_deap}
\title{Get DEAP ontology URL}
\usage{
get_url_deap(
  nodes,
  order,
  study,
  encode = TRUE,
  as_link = TRUE,
  link_type = "md",
  link_name = NULL
)
}
\arguments{
\item{nodes}{list of character vectors (length 2). A list of parent node &
end node pairs for each of the nodes that should be selected.}

\item{order}{character vector. The order of hierarchy levels to display in
the ontology tree from the following list:
\itemize{
\item \code{"study"}
\item \code{"domain"}
\item \code{"sub_domain"} (only for ABCD)
\item \code{"source"}
\item \code{"metric"} (only for ABCD)
\item \code{"atlas"} (only for ABCD)
}}

\item{study}{character. The name of the NBDC study (one of \code{"abcd"} or
\code{"hbcd"}).}

\item{encode}{logical. Whether or not to encode the URL (Default: \code{TRUE}). If
\code{FALSE}, argument \code{as_link} must also be \code{FALSE}.}

\item{as_link}{logical. Whether or not to return the URL formatted as a
hyperlink (Default: \code{TRUE}).}

\item{link_type}{character. The type of hyperlink to return (one of \code{"md"},
the default, or \code{"html"}).}

\item{link_name}{character. The name to use for the hyperlink (Default:
\code{NULL}, i.e., the URL is used as name)}
}
\value{
character. The URL or hyperlink for the selected node(s) and study.
}
\description{
Retrieves the URL for a position in the DEAP ontology tree. Allows to specify
a hierarchy order (including removing hierarchy levels) and a single or
multiple end node(s) that will show up as selected (and filter the data
dictionary table) when the URL is opened.
}
\examples{
NBDCtools:::get_url_deap(
  nodes = list(c("Standard Variables", "General")),
  order = c("study", "domain", "sub_domain", "source", "metric", "atlas"),
  study = "abcd",
  link_name = "ABCD 'Standard Variables' (standard order; Markdown link)"
)
NBDCtools:::get_url_deap(
  nodes = list(
    c("Imaging", "Desikan"),
    c("Imaging", "Destrieux")
  ),
  order = c("study", "domain", "atlas", "sub_domain", "metric"),
  study = "abcd",
  link_type = "html",
  link_name = "Desikan & Destrieux atlases (custom order; HTML link)"
)
NBDCtools:::get_url_deap(
  nodes = list(c("BioSpecimens", "Biological Mother")),
  order = c("study", "domain", "source"),
  study = "hbcd",
  as_link = FALSE # just return the URL, not a link
)
}
\keyword{internal}
