% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url_deap_table}
\alias{get_url_deap_table}
\alias{get_url_deap_table_abcd}
\alias{get_url_deap_table_hbcd}
\title{Get DEAP ontology URL for a specific table}
\usage{
get_url_deap_table(
  table,
  study,
  as_link = TRUE,
  link_type = "md",
  link_name = "name"
)

get_url_deap_table_abcd(...)

get_url_deap_table_hbcd(...)
}
\arguments{
\item{table}{character. The table name (\code{table_name} in the data dictionary).}

\item{study}{character. The name of the NBDC study (one of \code{"abcd"} or
\code{"hbcd"}).}

\item{as_link}{logical. Whether or not to return the URL formatted as a
hyperlink (Default: \code{TRUE}).}

\item{link_type}{character. The type of hyperlink to return (one of \code{"md"},
the default, or \code{"html"}).}

\item{link_name}{character. The name to use for the hyperlink (one of
\code{"name"}, i.e., use the table name, or \code{"label"}, i.e., use the table
label).}

\item{...}{Additional arguments passed to the underlying
\code{get_url_deap_table()} function.}
}
\value{
character. The URL or hyperlink for the specified table.
}
\description{
Wrapper around \code{get_url_deap()} to get the URL to open a specific table in
the DEAP ontology tree.

In addition to the main \code{get_url_deap_table()} function, there are two
study-specific variations:
\itemize{
\item \code{get_url_deap_table_abcd()}: for the ABCD study.
\item \code{get_url_deap_table_hbcd()}: for the HBCD study.
}

They have the same arguments as the \code{get_url_deap_table()} function, except
that the \code{study} argument is set to the respective study by default, and
should not be set by the user.
}
\examples{
NBDCtools:::get_url_deap_table("ab_g_dyn", "abcd")

NBDCtools:::get_url_deap_table("par_visit_data", "hbcd", link_name = "label")

NBDCtools:::get_url_deap_table_abcd("ab_g_dyn")

NBDCtools:::get_url_deap_table_hbcd("par_visit_data", link_name = "label")
}
\keyword{internal}
