% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_by_identifier}
\alias{join_by_identifier}
\title{Join by identifier set}
\usage{
join_by_identifier(
  dir_data,
  dd,
  identifiers,
  format = "parquet",
  shadow = FALSE
)
}
\arguments{
\item{dir_data}{character. Path to the directory with the data files in
\code{.parquet} or \code{.tsv} format.}

\item{dd}{tibble. Data frame with the data dictionary.}

\item{identifiers}{character (vector). Identifier column(s).}

\item{format}{character. Data format (One of \code{"parquet"} or \code{"tsv"}; default:
\code{"parquet"}).}

\item{shadow}{logical. Whether to join the shadow matrix
instead of the data table (default: \code{FALSE}).}
}
\value{
A tibble with the joined variables for the given (set of) identifier
column(s).
}
\description{
Internal helper for \code{\link[=join_tabulated]{join_tabulated()}} that joins requested variables from
all tables that have the given (set of) identifier column(s).
}
\keyword{internal}
