% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBDCtools-package.R
\docType{package}
\name{NBDCtools-package}
\alias{NBDCtools}
\alias{NBDCtools-package}
\title{NBDCtools: National Institutes of Health Brain Development Cohorts Data Hub Tools}
\description{
A suite of functions to work with data from the National Institutes of Health Brain Development Cohorts Data Hub. The package provides tools to create, clean, process, and filter datasets and associated metadata. These utilities are intended to simplify reproducible data-preparation for future research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://software.nbdc-datahub.org/NBDCtools/}
}

}
\author{
\strong{Maintainer}: Le Zhang \email{dairc.service@gmail.com} (\href{https://orcid.org/0009-0008-0205-2150}{ORCID})

Authors:
\itemize{
  \item Le Zhang \email{dairc.service@gmail.com} (\href{https://orcid.org/0009-0008-0205-2150}{ORCID})
  \item Janosch Linkersdoerfer \email{linkersdoerfer@gmail.com} (\href{https://orcid.org/0000-0002-1577-1233}{ORCID})
}

}
\keyword{internal}
