% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{read_file}
\alias{read_file}
\title{Read file}
\usage{
read_file(file, dd, format)
}
\arguments{
\item{file}{character. Path to the file.}

\item{dd}{tibble. Data frame with the data dictionary used to select columns
and determine the column types if reading from a \code{.tsv} file.}

\item{format}{character. Data format (One of \code{"parquet"} or \code{"tsv"})}
}
\value{
A tibble with the data/shadow matrix from the file.
}
\description{
Internal helper for \code{\link[=join_by_identifier]{join_by_identifier()}} that reads
data/shadow matrix for a given file in either
\code{.parquet} or \code{.tsv} format.
}
\keyword{internal}
