% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transf_factor}
\alias{transf_factor}
\title{Convert categorical columns to factor}
\usage{
transf_factor(data, study, release = "latest")
}
\arguments{
\item{data}{tibble. The data to be transformed. Columns are
expected to be in the data dictionary. If not, they will be skipped.}

\item{study}{character. NBDC study (One of \code{"abcd"} or \code{"hbcd"}.}

\item{release}{character. Release version (Default: \code{"latest"}).}
}
\value{
A tibble with the transformed data.
}
\description{
Based on the specifications in the data dictionary, transforms all
categorical columns to factor.
}
\examples{
\dontrun{
transf_factor(data, study = "abcd")
}
}
