% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_releases}
\alias{get_releases}
\alias{get_releases_abcd}
\alias{get_releases_hbcd}
\alias{get_latest_release}
\alias{get_latest_release_abcd}
\alias{get_latest_release_hbcd}
\title{Get release version number(s)}
\usage{
get_releases(study)

get_releases_abcd()

get_releases_hbcd()

get_latest_release(study)

get_latest_release_abcd()

get_latest_release_hbcd()
}
\arguments{
\item{study}{character. The study name. One of "abcd" or "hbcd".}
}
\value{
character. The latest release version number(s)
of the specified study.
}
\description{
These functions retrieve all release version number(s)
for a given study, or the latest release version number.
}
\examples{
\dontshow{if (requireNamespace("NBDCtoolsData", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_releases("abcd")
get_releases("hbcd")
get_latest_release("abcd")
get_latest_release("hbcd")
\dontshow{\}) # examplesIf}
}
