% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBDCtools-package.R
\docType{package}
\name{NBDCtools-package}
\alias{NBDCtools}
\alias{NBDCtools-package}
\title{NBDCtools: NBDCtools: NBDC Data Hub Tools}
\description{
This package provides functions to work with data from the NBDC Data Hub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://software.nbdc-datahub.org/NBDCtools/}
}

}
\author{
\strong{Maintainer}: Janosch Linkersdoerfer \email{dairc.service@gmail.com} (\href{https://orcid.org/0000-0002-1577-1233}{ORCID})

Authors:
\itemize{
  \item Le Zhang \email{lezhang100@gmail.com} (\href{https://orcid.org/0009-0008-0205-2150}{ORCID})
}

}
\keyword{internal}
